add errors

+++ backend/routes/posts.js
@@ -39,7 +39,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.memberData.memberId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -49,6 +50,10 @@ router.post(
           id: createdPost._id
         }
       });
+    }).catch( err => {
+      res.status(500).json({
+        message: 'Nie udało się dodać treśći'
+      })
     });
   }
 );
@@ -66,11 +71,21 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath, 
+      creator: req.memberData.memberId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    // console.log(post);
+    Post.updateOne({ _id: req.params.id, creator: req.memberData.memberId }, post).then(result => {
+      // console.log(result);
+      if ( result.nModified > 0 ) {
+        res.status(200).json({ message: "Update successful" });
+      } else {
+        res.status(401).json({ message: "No authorized" });
+      }
+    }).catch( err => {
+      res.status(500).json({
+        message: 'Nie udało się zaktualizować treśći'
+      })
     });
   }
 );
@@ -94,6 +109,10 @@ router.get("", (req, res, next) => {
         posts: fetchedPosts,
         maxPosts: count
       });
+    }).catch( err => {
+      res.status(500).json({
+        message: 'Nie udało się wczytać treśći'
+      })
     });
 });
 
@@ -104,13 +123,26 @@ router.get("/:id", (req, res, next) => {
     } else {
       res.status(404).json({ message: "Post not found!" });
     }
+  }).catch( err => {
+    res.status(500).json({
+      message: 'Nie znaleziono treści'
+    })
   });
 });
 
 router.delete("/:id", checkUser, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.memberData.memberId }).then(result => {
+    // console.log(result);
+    // res.status(200).json({ message: "Post deleted!" });
+    if ( result.n > 0 ) {
+      res.status(200).json({ message: "Usunięto" });
+    } else {
+      res.status(401).json({ message: "Nie można usunąć treśći" });
+    }
+  }).catch( err => {
+    res.status(500).json({
+      message: 'Nie można usunąć treśći'
+    })
   });
 });
